# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.AccountLimitView:
    properties:
      label:
        type: string
      limit_key:
        type: string
      resets_at:
        type: string
      status:
        type: string
      used_percent:
        type: number
      window_minutes:
        type: integer
    type: object
  models.AccountUsageView:
    properties:
      account_detail:
        type: string
      billing_label:
        type: string
      error:
        type: string
      extra_limits:
        items:
          $ref: '#/definitions/models.AccountLimitView'
        type: array
      extra_usage_label:
        type: string
      extra_usage_monthly_usd:
        type: number
      extra_usage_used_usd:
        type: number
      limits:
        items:
          $ref: '#/definitions/models.AccountLimitView'
        type: array
      plan_type:
        type: string
      primary_limit:
        $ref: '#/definitions/models.AccountLimitView'
      provider:
        type: string
      secondary_limit:
        $ref: '#/definitions/models.AccountLimitView'
      status_label:
        type: string
      updated_at:
        type: string
    type: object
...
  models.AnalyticsUsageViewModel:
    properties:
      account_limits:
        items:
          $ref: '#/definitions/models.AccountUsageView'
        type: array
      daily_usage:
        items:
          $ref: '#/definitions/models.DailyUsagePoint'
        type: array
      daily_usage_by_model:
        items:
          $ref: '#/definitions/models.DailyUsagePoint'
        type: array
      errors:
        items:
          type: string
        type: array
      last_updated_at:
        type: string
      model_breakdown:
        items:
          $ref: '#/definitions/models.ModelUsagePoint'
        type: array
      totals:
        $ref: '#/definitions/models.UsageTotals'
      usage_rate:
        items:
          $ref: '#/definitions/models.UsageRatePoint'
        type: array
      usage_rate_by_model:
        items:
          $ref: '#/definitions/models.UsageRatePoint'
        type: array
    type: object
...
  models.DailyUsagePoint:
    properties:
      cached_input_tokens:
        type: integer
      call_count:
        type: integer
      cost_usd:
        type: number
      input_tokens:
        type: integer
      model:
        type: string
      output_tokens:
        type: integer
      period:
        type: string
      provider:
        type: string
      total_tokens:
        type: integer
    type: object
...
  models.ModelUsagePoint:
    properties:
      cached_input_tokens:
        type: integer
      call_count:
        type: integer
      cost_usd:
        type: number
      input_tokens:
        type: integer
      model:
        type: string
      output_tokens:
        type: integer
      percent:
        type: number
      provider:
        type: string
      reasoning_output_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
...
  models.SkillAgentUsageAgent:
    properties:
      agent_id:
        type: string
      agent_name:
        type: string
    type: object
  models.SkillAgentUsageCell:
    properties:
      activity_count:
        type: integer
      agent_id:
        type: string
      agent_name:
        type: string
      edited_count:
        type: integer
      loaded_count:
        type: integer
      selected_count:
        type: integer
      skill_handle:
        type: string
      viewed_count:
        type: integer
    type: object
  models.SkillAgentUsageHeatmap:
    properties:
      agents:
        items:
          $ref: '#/definitions/models.SkillAgentUsageAgent'
        type: array
      cells:
        items:
          $ref: '#/definitions/models.SkillAgentUsageCell'
        type: array
      skills:
        items:
          type: string
        type: array
    type: object
  models.SkillAnalyticsDashboard:
    properties:
      agent_usage:
        $ref: '#/definitions/models.SkillAgentUsageHeatmap'
      follow_through:
        items:
          $ref: '#/definitions/models.SkillFollowThroughMetric'
        type: array
      top_skills:
        items:
          $ref: '#/definitions/models.SkillAnalyticsSkillMetric'
        type: array
      underused:
        items:
          $ref: '#/definitions/models.UnderusedSkillMetric'
        type: array
    type: object
  models.SkillAnalyticsSkillMetric:
    properties:
      activity_count:
        type: integer
      edited_count:
        type: integer
      follow_through_rate:
        type: number
      last_activity:
        type: string
      loaded_count:
        type: integer
      selected_count:
        type: integer
      skill_handle:
        type: string
      skill_scope:
        type: string
      viewed_count:
        type: integer
    type: object
  models.SkillFollowThroughMetric:
    properties:
      ignored_count:
        type: integer
      loaded_or_viewed_count:
        type: integer
      selected_count:
        type: integer
      skill_handle:
        type: string
      skill_scope:
        type: string
    type: object
  models.UnderusedSkillMetric:
    properties:
      activity_count:
        type: integer
      always_use:
        type: boolean
      edited_count:
        type: integer
      enabled:
        type: boolean
      last_activity:
        type: string
      loaded_count:
        type: integer
      selected_count:
        type: integer
      skill_handle:
        type: string
      skill_scope:
        type: string
      viewed_count:
        type: integer
    type: object
  models.UsageRatePoint:
    properties:
      call_count:
        type: integer
      model:
        type: string
      period:
        type: string
      provider:
        type: string
      total_tokens:
        type: integer
    type: object
  models.UsageTotals:
    properties:
      cache_creation_input_tokens:
        type: integer
      cache_read_input_tokens:
        type: integer
      cached_input_tokens:
        type: integer
      call_count:
        type: integer
      cost_available:
        type: boolean
      cost_usd:
        type: number
      input_tokens:
        type: integer
      output_tokens:
        type: integer
      reasoning_output_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
...
  repository.AgentUsage:
    properties:
      agentID:
        type: string
      agentName:
        type: string
      executionCount:
        type: integer
      failureCount:
        type: integer
      projectID:
        type: string
      projectName:
        type: string
      successCount:
        type: integer
    type: object
  repository.AvgExecutionTime:
    properties:
      avgMs:
        format: float64
        type: number
      count:
        type: integer
      id:
        type: string
      maxMs:
        format: int64
        type: integer
      minMs:
        format: int64
        type: integer
      name:
        type: string
    type: object
  repository.ExecutionTrend:
    properties:
      count:
        type: integer
      hour:
        type: integer
    type: object
  repository.FailedTaskPattern:
    properties:
      failureCount:
        type: integer
      lastError:
        type: string
      lastFailedAt:
        type: string
      taskID:
        type: string
      taskTitle:
        type: string
    type: object
  repository.SuccessFailureRate:
    properties:
      failureCount:
        type: integer
      period:
        type: string
      successCount:
        type: integer
      successRate:
        format: float64
        type: number
      totalCount:
        type: integer
    type: object
  repository.TaskFrequency:
    properties:
      executionCount:
        type: integer
      lastExecutedAt:
        type: string
      taskID:
        type: string
      taskTitle:
        type: string
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@openvibely.dev
    name: API Support
    url: https://github.com/openvibely/openvibely
  description: |-
    REST API for OpenVibely - AI-powered task scheduling and management
    This API provides endpoints for managing projects, tasks, and chat interactions with AI agents.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OpenVibely API
  version: "1.0"
paths:
  /api/analytics/agent-usage-by-project:
    get:
      description: Returns model usage, success count, and failure count grouped by
        project.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Model usage breakdown
          schema:
            items:
              $ref: '#/definitions/repository.AgentUsage'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get model usage by project
      tags:
      - analytics
  /api/analytics/avg-execution-time-by-agent:
    get:
      description: Returns average execution durations grouped by configured model.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Average execution time by model
          schema:
            items:
              $ref: '#/definitions/repository.AvgExecutionTime'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get average execution time by model
      tags:
      - analytics
  /api/analytics/avg-execution-time-by-task:
    get:
      description: Returns average task execution durations for completed executions.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - default: 10
        description: Maximum number of tasks to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Average execution time by task
          schema:
            items:
              $ref: '#/definitions/repository.AvgExecutionTime'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get average execution time by task
      tags:
      - analytics
  /api/analytics/execution-trends-by-hour:
    get:
      description: Returns execution counts grouped by hour-of-day.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - description: Optional start datetime filter
        in: query
        name: date_from
        type: string
      - description: Optional end datetime filter
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Execution counts by hour
          schema:
            items:
              $ref: '#/definitions/repository.ExecutionTrend'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get execution trends by hour
      tags:
      - analytics
  /api/analytics/failed-task-patterns:
    get:
      description: Returns tasks with repeated failures, grouped by last observed
        error.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - default: 10
        description: Maximum number of patterns to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Failed task patterns
          schema:
            items:
              $ref: '#/definitions/repository.FailedTaskPattern'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get failed task patterns
      tags:
      - analytics
  /api/analytics/most-frequent-tasks:
    get:
      description: Returns tasks ordered by execution count.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - default: 10
        description: Maximum number of tasks to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Most frequently executed tasks
          schema:
            items:
              $ref: '#/definitions/repository.TaskFrequency'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get most frequent tasks
      tags:
      - analytics
  /api/analytics/skills:
    get:
      description: Returns top skills, selection follow-through, agent heatmap, and
        underused skill metrics.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - default: 30d
        description: 'Convenience range: 7d, 30d, 90d, all'
        in: query
        name: range
        type: string
      - description: Agent ID filter
        in: query
        name: agent_id
        type: string
      - description: Surface filter
        in: query
        name: surface
        type: string
      - description: Skill scope filter
        in: query
        name: skill_scope
        type: string
      - description: Event type filter
        in: query
        name: event_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Skill analytics
          schema:
            $ref: '#/definitions/models.SkillAnalyticsDashboard'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get skill analytics
      tags:
      - analytics
  /api/analytics/success-failure-rates:
    get:
      description: Returns execution success/failure rates grouped by day, week, or
        month.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - default: day
        description: 'Grouping period: day, week, or month'
        in: query
        name: group_by
        type: string
      - description: Optional start datetime filter
        in: query
        name: date_from
        type: string
      - description: Optional end datetime filter
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success/failure rates
          schema:
            items:
              $ref: '#/definitions/repository.SuccessFailureRate'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get success/failure rates
      tags:
      - analytics
  /api/analytics/usage:
    get:
      description: Returns token/cache/reasoning/cost totals, daily usage, usage rate,
        model breakdowns, and account limit snapshots.
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - description: Provider filter
        in: query
        name: provider
        type: string
      - default: 30d
        description: 'Convenience range: 7d, 30d, month, all'
        in: query
        name: range
        type: string
      - default: day
        description: 'Usage rate grouping: hour, day, week, month'
        in: query
        name: group_by
        type: string
      - description: Optional start datetime filter
        in: query
        name: date_from
        type: string
      - description: Optional end datetime filter
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usage analytics
          schema:
            $ref: '#/definitions/models.AnalyticsUsageViewModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get LLM usage analytics
      tags:
      - analytics
...
schemes:
- http
- https
swagger: "2.0"
...